updater
Enum Task

pub(crate) enum Task {
    Upload,
    IncrementalUpload {
        version: String,
    },
    DryRun {
        single_file: Option<String>,
    },
    OptIntoAll,
    Download {
        minimal: bool,
        dont_delete: bool,
        dl_from_local: bool,
        version: String,
    },
}

Variants

Upload

Synchronize the source-of-truth in S3 with data in the local directory. Based on current permissions, only Dustin runs this.

IncrementalUpload

This uploads to S3 from cloud VMs that import maps. This never deletes files from S3, only updates or creates ne ones.

Fields

version: String

Upload data to a temporary version managed by the cloud scripts.

DryRun

Just compare data in the current directory with the manifest, and describe any new, deleted, or modified files.

Fields

single_file: Option<String>

Just check if one file has changed.

OptIntoAll

Print the JSON list of all possible city data packs to download. You can write this output to data/player/data.json, then download everything.

Download

Synchronize the local data directory with the source-of-truth in S3.

Fields

minimal: bool

The Github Actions build uses this to include only a few files for the release to be usable. People can use the UI to open another map and download more data.

dont_delete: bool

Only update files from the manifest. Leave extra files alone.

dl_from_local: bool

Only useful for Dustin. “Download” from my local S3 source-of-truth, not from the network.

version: String

Download data tied to a named release. See https://a-b-street.github.io/docs/tech/dev/data.html.